+++ UICatalog/Scenarios/CsvEditor.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Data;
 using Terminal.Gui;
+using Terminal.Gui.Trees;
 using System.Linq;
 using System.Globalization;
 using System.IO;
