+++ UnitTests/TreeViewTests.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Terminal.Gui;
+using Terminal.Gui.Trees;
 using Xunit;
 
 namespace Terminal.Gui.Views {
